# apps/backend/swagger.yml
swagger: "2.0"
info:
  description: "This is a simple API for managing todo items."
  version: "1.0.0"
  title: "Todo API"
host: "localhost:5000"
basePath: "/"
schemes:
- "http"
paths:
  /todos:
    get:
      summary: "Lists all the todos"
      operationId: "getTodos"
      produces:
      - "application/json"
      responses:
        200:
          description: "An array of todos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Todo"
    post:
      summary: "Create a new todo"
      operationId: "addTodo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Todo object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Todo"
      responses:
        200:
          description: "New todo created"
          schema:
            $ref: "#/definitions/Todo"
  /todos/{id}:
    get:
      summary: "Find todo by ID"
      description: "Returns a single todo"
      operationId: "getTodoById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of todo to return"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Todo"
        404:
          description: "Todo not found"
    put:
      summary: "Updates a todo"
      description: "Updates an existing todo"
      operationId: "updateTodo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of todo that needs to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated todo object"
        required: true
        schema:
          $ref: "#/definitions/Todo"
      responses:
        200:
          description: "Todo updated"
          schema:
            $ref: "#/definitions/Todo"
        404:
          description: "Todo not found"
    delete:
      summary: "Deletes a todo"
      description: "Deletes an existing todo"
      operationId: "deleteTodo"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the todo that needs to be deleted"
        required: true
        type: "string"
      responses:
        200:
          description: "Todo deleted"
        404:
          description: "Todo not found"
definitions:
  Todo:
    type: "object"
    required:
    - "title"
    properties:
      _id:
        type: "string"
        description: "The auto-generated id of the todo"
      title:
        type: "string"
        description: "The title of the todo item"
      completed:
        type: "boolean"
        description: "The completion status of the todo item"
    example:
      _id: "d5fE_asz"
      title: "Walk the dog"
      completed: false